app_docker_env/requirements.txt
Flask==2.0.2
redis==4.1.1
mysql-connector-python==8.0.27
